# service/userService.go
package service

import (
	"IM_Project/models"
	"IM_Project/utils"
	"fmt"
	"github.com/gin-gonic/gin"
	"github.com/gorilla/websocket"
	"log"
	"math/rand"
	"net/http"
	"strconv"
	"time"
)

func GetUserList(c *gin.Context) {
	data := make([]*models.UserBasic, 10)
	data = models.GetUserList()
	c.JSON(http.StatusOK, gin.H{
		"message": data,
	})
}

// FindUserByNameAndPwd
// @Summary 通过用户名与密码查找用户
// @Tags 用户模块
// @param name query string false "用户名"
// @param password query string false "密码"
// @Success 200 {string} json{"code","message"}
// @Router /user/findUserByNameAndPwd [post]
func FindUserByNameAndPwd(c *gin.Context) {
	name := c.Request.FormValue("name")
	password := c.Request.FormValue("password")
	user := models.FindUserByName(name)
	if user.Name == "" {
		utils.Error(c, -1, "该用户不存在")
		return
	}

	flag := utils.MakePassword(password, user.Salt)
	pwd := models.FindUserByNameAndPwd(name, flag)
	if pwd.Name == "" {
		utils.Error(c, -1, "密码不正确")
		return
	}

	utils.Success(c, 1, "登录成功", user)
}

// CreateUser
// @Summary 新增用户
// @Tags 用户模块
// @param name query string false "用户名"
// @param password query string false "密码"
// @param confirmPassword query string false "确认密码"
// @Success 200 {string} json{"code","message"}
// @Router /user/createUser [get]
func CreateUser(c *gin.Context) {

	user := models.UserBasic{}
	user.Name = c.Request.FormValue("name")
	password := c.Request.FormValue("password")
	confirmPassword := c.Request.FormValue("confirmPassword")
	fmt.Println(user.Name, "  >>>>>>>>>>>  ", password, confirmPassword)
	salt := fmt.Sprintf("%06d", rand.Int31())

	if user.Name == "" || password == "" || confirmPassword == "" {
		utils.Error(c, -1, "用户名或密码不能为空！")
		return
	}
	data := models.FindUserByName(user.Name)
	if data.Name != "" {
		utils.Error(c, -1, "用户名已注册！")
		return
	}
	if password != confirmPassword {
		utils.Error(c, -1, "两次密码不一致！")
		return
	}

	user.PassWord = utils.MakePassword(password, salt)
	user.Salt = salt
	user.LoginTime = time.Now()
	user.LoginOutTime = time.Now()
	user.HeartbeatTime = time.Now()
	models.CreateUser(user)
	utils.Success(c, 1, "新增用户成功！", user)
}

// DeleteUser
// @Summary 删除用户
// @Tags 用户模块
// @param id query string false "id"
// @Success 200 {string} json{"code","message"}
// @Router /user/deleteUser [post]
func DeleteUser(c *gin.Context) {
	user := models.UserBasic{}
	id, _ := strconv.Atoi(c.Query("id"))
	user.ID = uint(id)
	models.DeleteUser(user)
	utils.Success(c, 1, "删除用户成功！", user)

}

// UpdateUser
// @Summary 修改用户
// @Tags 用户模块
// @param id formData string false "id"
// @param name formData string false "name"
// @param password formData string false "password"
// @param phone formData string false "phone"
// @param email formData string false "email"
// @Success 200 {string} json{"code","message"}
// @Router /user/updateUser [post]
func UpdateUser(c *gin.Context) {
	user := models.UserBasic{}
	id, _ := strconv.Atoi(c.PostForm("id"))
	user.ID = uint(id)
	user.Name = c.PostForm("name")
	user.PassWord = c.PostForm("password")
	user.Phone = c.PostForm("phone")
	user.Avatar = c.PostForm("icon")
	user.Email = c.PostForm("email")

	models.UpdateUser(user)
	user = models.FindByID(user.ID)
	utils.Success(c, 1, "修改用户成功！", user)
}

//防止跨域站点伪造请求
var upGrader = websocket.Upgrader{
	CheckOrigin: func(r *http.Request) bool {
		return true
	},
}

func SendMsg(c *gin.Context) {
	ws, err := upGrader.Upgrade(c.Writer, c.Request, nil)
	if err != nil {
		fmt.Println(err)
		return
	}
	defer func(ws *websocket.Conn) {
		err = ws.Close()
		if err != nil {
			fmt.Println(err)
		}
	}(ws)
	MsgHandler(c, ws)
}

func MsgHandler(c *gin.Context, ws *websocket.Conn) {
	for {
		msg, err := utils.Subscribe(c, utils.PublishKey)
		if err != nil {
			fmt.Println(" MsgHandler 发送失败", err)
		}

		tm := time.Now().Format("2006-01-02 15:04:05")
		m := fmt.Sprintf("[ws][%s]:%s", tm, msg)
		err = ws.WriteMessage(1, []byte(m))
		if err != nil {
			log.Fatalln(err)
		}
	}
}

func SendUserMsg(c *gin.Context) {
	models.Chat(c.Writer, c.Request)
}

func AddFriend(c *gin.Context) {
	userId, _ := strconv.Atoi(c.Request.FormValue("userId"))
	targetName := c.Request.FormValue("targetName")
	code, msg := models.AddFriend(uint(userId), targetName)
	if code == 0 {
		utils.RespOK(c.Writer, code, msg)
	} else {
		utils.RespFail(c.Writer, msg)
	}
}

func SearchFriends(c *gin.Context) {
	id, _ := strconv.Atoi(c.Request.FormValue("userId"))
	users := models.SearchFriend(uint(id))
	utils.RespOKList(c.Writer, users, len(users))
}

//func RedisMsg(c *gin.Context) {
//	userIdA, _ := strconv.Atoi(c.PostForm("userIdA"))
//	userIdB, _ := strconv.Atoi(c.PostForm("userIdB"))
//	start, _ := strconv.Atoi(c.PostForm("start"))
//	end, _ := strconv.Atoi(c.PostForm("end"))
//	isRev, _ := strconv.ParseBool(c.PostForm("isRev"))
//	res := models.RedisMsg(int64(userIdA), int64(userIdB), int64(start), int64(end), isRev)
//	utils.RespOKList(c.Writer, "ok", res)
//}